+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/PostgreObject.java
@@ -35,6 +35,6 @@ public interface PostgreObject extends DBSObject {
     /**
      * OID
      */
-    int getObjectId();
+    long getObjectId();
 
 }
